警告つぶし。

+++ NVEncCore/rgy_input.cpp
@@ -146,9 +146,9 @@ int RGYConvertCSP::run(int interlaced, void **dst, const void **src, int width, 
     }
     m_csp->func[interlaced](dst, src,
         width, src_y_pitch_byte, src_uv_pitch_byte, dst_y_pitch_byte,
-        height, dst_height, 0, m_th.size()+1, crop);
+        height, dst_height, 0, (int)m_th.size()+1, crop);
     if (m_th.size() > 0) {
-        WaitForMultipleObjects(m_heFinCopy.size(), m_heFinCopy.data(), TRUE, INFINITE);
+        WaitForMultipleObjects((DWORD)m_heFinCopy.size(), m_heFinCopy.data(), TRUE, INFINITE);
     }
     return 0;
 }
